fix: address review findings for ideation page
- Remove eslint-disable comment and ccstate-react/experimental usage,
  replace with React useState for local tab state
- Extract use case data to separate zero-ideation-data.ts file
- Fix broken tests that relied on removed hardcoded prompt titles
- Fix lint violation (toHaveLength instead of toBe for length check)
- Remove unused type exports flagged by knip
